abstract sourcemoderation logic
